refactor(nicolive): 失敗オブジェクトを移動
