+++ app/components/nicolive-area/NicoliveArea.vue.ts
@@ -1,9 +1,7 @@
 import Vue from 'vue';
 import { Component } from 'vue-property-decorator';
 import { Inject } from 'util/injector';
-import { NicoliveProgramService, NicoliveProgramServiceFailure } from 'services/nicolive-program/nicolive-program';
-import { remote } from 'electron';
-import { $t } from 'services/i18n';
+import { NicoliveProgramService } from 'services/nicolive-program/nicolive-program';
 
 import ProgramDescription from './ProgramDescription.vue';
 import CommentViewer from './CommentViewer.vue';
@@ -14,6 +12,7 @@ import ToolBar from './ToolBar.vue';
 import TopNav from './TopNav.vue';
 import ControlsArrow from '../../../media/images/controls-arrow-vertical.svg';
 import AreaSwitcher from './AreaSwitcher.vue';
+import { NicoliveFailure, openErrorDialogFromFailure } from 'services/nicolive-program/NicoliveFailure';
 
 @Component({
   components: {
@@ -75,8 +74,8 @@ export default class NicolivePanelRoot extends Vue {
       this.isFetching = true;
       await this.nicoliveProgramService.fetchProgram();
     } catch (caught) {
-      if (caught instanceof NicoliveProgramServiceFailure) {
-        await NicoliveProgramService.openErrorDialogFromFailure(caught);
+      if (caught instanceof NicoliveFailure) {
+        await openErrorDialogFromFailure(caught);
       } else {
         throw caught;
       }
